Change creditHour type from int to float in GpaRecord structure, after a test failed.

+++ include/gpa.h
@@ -15,7 +15,7 @@ from running to avoid duplicate definition errors.
 // Structure for storing credit hour and grade point of each course
 typedef struct
 {
-    int creditHour;
+    float creditHour;
     float gradePoint;
 
 } GpaRecord;
